refactor: Separated layer of popup and layer of form, used common DialogForm
